+++ src/components/CopyUiProvider/index.ts
@@ -1,6 +1,7 @@
-// CopyUiProvider from copy-ui @ 2025-11-17.
+// CopyUiProvider from copy-ui @ 2025-11-22.
 
 export { CopyUiProvider } from "./copy-ui-provider";
 export type { ColorName, ColorNo, Theme } from "./theme";
 export { useJss } from "./use-jss";
+export { useMode, useSetMode } from "./use-mode";
 export { useTheme } from "./use-theme";